feature : 주문 도메인로직 구현
- 상품들의 재고를 확인하고 재고를 업데이트, 총 주문 금액을 계산하는 컴포넌트 구현 (orderLineAggregator, OrderLineAllocator)
- 주문의 저장하고 상품의 총 가격을 결제하고, 사용자의 포인트를 감소하는 컴포넌트 구현
(OrderCashier)

+++ core/service/user-service/src/test/java/com/loopers/core/service/user/UserPointQueryServiceIntegrationTest.java
@@ -75,7 +75,7 @@ class UserPointQueryServiceIntegrationTest extends IntegrationTest {
                 SoftAssertions.assertSoftly(softly -> {
                     softly.assertThat(userPoint).isNotNull();
                     softly.assertThat(userPoint.getUserId().value()).isEqualTo(savedUser.getUserId().value());
-                    softly.assertThat(userPoint.getBalance().value()).isEqualTo(0);
+                    softly.assertThat(userPoint.getBalance().value().intValue()).isEqualTo(0);
                 });
             }
         }
